feat: altera a função de deletar para Missoes e Ninjas, agora a função valida o ID antes de deletar

+++ cadastroNinjas/src/main/java/dev/java10x/cadastroNinjas/Missoes/MissoesService.java
@@ -52,8 +52,12 @@ public class MissoesService{
     }
 
     // Deletar Missoes(DELETE)
-    public void deletarMissoes(Long id){
-        missoesRepository.deleteById(id);
+    public Long deletarMissoes(Long id){
+        if (missoesRepository.existsById(id)) {
+            missoesRepository.deleteById(id);
+            return id;
+        }
+        return null;
     }
 }
